refactor(static-embeddings): move content of word.py to embedding.py
