+++ embeddings/embedding/auto_flair.py
@@ -6,18 +6,15 @@ from embeddings.embedding.flair_embedding import (
     FlairTransformerDocumentEmbedding,
     FlairTransformerWordEmbedding,
 )
-from embeddings.embedding.static.word import AutoStaticDocumentEmbedding, AutoStaticWordEmbedding
+from embeddings.embedding.static.embedding import (
+    AutoStaticDocumentEmbedding,
+    AutoStaticWordEmbedding,
+)
 
 
 class AutoFlairEmbedding(ABC):
     @staticmethod
     @abstractmethod
-    def from_hub(repo_id: str, **kwargs: Any) -> FlairEmbedding:
-        pass
-
-
-class AutoFlairWordEmbedding(AutoFlairEmbedding):
-    @staticmethod
     def from_hub(repo_id: str, **kwargs: Any) -> FlairEmbedding:
         """Loads an embedding model from hugging face hub, if the model is compatible with
         Transformers or the current library.
@@ -25,7 +22,12 @@ class AutoFlairWordEmbedding(AutoFlairEmbedding):
         In case of StaticWordEmbedding, a default config is used during initialisation. If you
         rather want to specify custom config, use the AutoStaticWordEmbedding.from_config function.
         """
+        pass
 
+
+class AutoFlairWordEmbedding(AutoFlairEmbedding):
+    @staticmethod
+    def from_hub(repo_id: str, **kwargs: Any) -> FlairEmbedding:
         try:
             return FlairTransformerWordEmbedding(repo_id, **kwargs)
         except EnvironmentError:
@@ -35,17 +37,9 @@ class AutoFlairWordEmbedding(AutoFlairEmbedding):
 class AutoFlairDocumentEmbedding(AutoFlairEmbedding):
     @staticmethod
     def from_hub(repo_id: str, **kwargs: Any) -> FlairEmbedding:
-        """Loads an embedding model from hugging face hub, if the model is compatible with
-        Transformers or the current library.
-
-        In case of StaticWordEmbedding, a default config is used during initialisation. If you
-        rather want to specify custom config, use the AutoFlairDocumentEmbedding.from_config
-        function.
-
-        Additionally, in case of StaticWordEmbedding, pooling ("mean" by default) of such embeddings
-         is performed to obtain a document's embedding.
+        """In case of StaticWordEmbedding, mean pooling on such embeddings is performed to obtain a
+        document's embedding.
         """
-
         try:
             return FlairTransformerDocumentEmbedding(repo_id, **kwargs)
         except EnvironmentError:
